requirements.txt
Django==1.6.6
Markdown==2.4
Pillow==2.3.0
PyJWT==0.3.0
Pygments==1.6
South==1.0.1
boto==2.34.0
django-filter==0.7
djangorestframework==2.4.3
djangorestframework-jwt==1.0.2
openpyxl==1.7.0
psycopg2==2.4.5
